requirements.txt
beautifulsoup4
newspaper3k
transformers
requests
oauth2client
gspread